refactor(#2819): keep over-threshold files from growing (campsite rule)
Extract the flight sub-phase value table + emitter into a new
cqlite-flight/src/obs_subphase.rs sibling (obs.rs back to its pre-change
size), drop the unneeded cold-path producer_thread propagation in the
13.5k-line merge/mod.rs (flight do_get uses the warm reader path only), and
revert the catalog.rs const-doc growth (the under-threshold
operator_docs_annotations.rs carries the authoritative sub-phase annotation,
regenerated into the committed metrics docs). Move StreamSubPhaseEmitter
call site to crate::obs_subphase.
