+++ cqlite-flight/src/lib.rs
@@ -18,6 +18,7 @@ pub mod cancel;
 pub mod filter;
 pub mod obs;
 pub mod obs_abort;
+pub mod obs_subphase;
 pub mod pathsafe;
 pub mod point_read;
 pub mod producer;
